app/requirements.txt
python-consul2==0.1.5
Flask==2.1.2
Flask_Migrate==3.1.0
Flask_OAuthlib==0.9.6
Flask_SQLAlchemy==2.5.1
requests==2.27.1
psycopg2-binary==2.9.3
Flask-APScheduler==1.12.3